+++ core/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 All notable changes to this project will be documented in this file. This project make usage of the [Yii Versioning Strategy](https://github.com/yiisoft/yii2/blob/master/docs/internals/versions.md). In order to read more about upgrading and BC breaks have a look at the [UPGRADE Document](UPGRADE.md).
 
+## 1.0.14 (in progress)
+
 ## 1.0.13 (30. October 2018)
 
 + [#1869](https://github.com/luyadev/luya/issues/1869) Fix bug with word highlight in string helper.